refactor: 상품 삭제 시 position 대신 id를 사용하도록 수정

+++ app/src/main/java/woowacourse/shopping/data/DefaultCartRepository.kt
@@ -15,7 +15,7 @@ class DefaultCartRepository(
     override suspend fun getAllCartProducts(): List<Product> =
         cartProductDao.getAll().map { it.toDomain() }
 
-    override suspend fun deleteCartProduct(id: Int) {
-        cartProductDao.delete(id.toLong())
+    override suspend fun deleteCartProduct(id: Long) {
+        cartProductDao.delete(id)
     }
 }
